+++ R/intData_utils.r
@@ -31,7 +31,7 @@
 #' # Define grouping variable for microdata aggregation
 #' credit_agrby <- paste(CreditCard_microdata$Name, CreditCard_microdata$Month, sep = "_")
 #' 
-#' # Obtain latent variables inherent to the macrodata (standardized to [-1,1])
+#' # Obtain latent variables inherent to the macrodata (normalized to [-1,1])
 #' credit_card_U <- get_latent_var(microdata = CreditCard_microdata[,3:7], 
 #'                                 macrodata = CreditCard_min_max, 
 #'                                 agrby = credit_agrby, 
@@ -114,7 +114,13 @@ get_latent_var <- function(microdata,
 #' @param p Number of variables.
 #' @param estimate.DistParam Logical parameter indicating if estimation of the parameters of the latent distributions should be performed. Can only be set to TRUE if \code{LatentCase="General"}.
 #' The default is \code{FALSE}.
-#' @return A list with the parameters of the latent variables.
+#' @return A list composed by:
+#' \item{\code{LatentParam}}{A list with the parameters of the latent variables distribution.}
+#' \item{\code{TriangParam}}{The mode of the triangular distribution.}
+#' \item{\code{BetaParam.a}}{The alpha parameter of the Beta distribution.}
+#' \item{\code{BetaParam.b}}{The beta parameter of the Beta distribution.}
+#' \item{\code{LatentCase}}{The case of the latent variables.}
+#' \item{\code{LatentDist}}{The distribution of the latent variables.}
 #' @export
 #' @examples
 #' data(creditcard)
@@ -124,7 +130,7 @@ get_latent_var <- function(microdata,
 #' # Define grouping variable for microdata aggregation
 #' credit_agrby <- paste(CreditCard_microdata$Name, CreditCard_microdata$Month, sep = "_")
 #' 
-#' # Obtain latent variables inherent to the macrodata (standardized to [-1,1])
+#' # Obtain latent variables inherent to the macrodata (normalized to [-1,1])
 #' credit_card_U <- get_latent_var(microdata = CreditCard_microdata[,3:7], 
 #'                                 macrodata = CreditCard_min_max, 
 #'                                 agrby = credit_agrby, 
@@ -385,6 +391,12 @@ CalE.beta.beta <- function(a1,b1,a2,b2){
 #' @return  Value
 #' @keywords internal
 CalE.beta.kde <- function(micro,a1,b1){
+    # A constant sample represents a degenerate distribution, for which the
+    # cross-moment is available directly.
+    if(length(unique(micro)) == 1L){
+        return(mean(micro)*(2*a1/(a1+b1)-1))
+    }
+
     fit3 <- kde1d::kde1d(micro) # estimate density
     integrandBetaBeta.kde <- function(x,fit3,a=a1,b=b1) {qbeta(x,a,b)*kde1d::qkde1d(x, fit3)}
 
@@ -400,6 +412,12 @@ CalE.beta.kde <- function(micro,a1,b1){
 #' @return  Value
 #' @keywords internal
 CalE.kde.kde <- function(micro1,micro2){
+    # A local-quadratic KDE is not identifiable for a constant sample. Its
+    # quantile is constant, so the cross-moment is the product of the means.
+    if(length(unique(micro1)) == 1L || length(unique(micro2)) == 1L){
+        return(mean(micro1)*mean(micro2))
+    }
+
     fit3 <- kde1d::kde1d(micro1) # estimate density
     fit4 <- kde1d::kde1d(micro2) # estimate density
     integrand.kde.kde <- function(x,fit3=fit3,fit4=fit4) {kde1d::qkde1d(x, fit3)*kde1d::qkde1d(x, fit4)}