+++ R/plots_Shapley.r
@@ -560,6 +560,7 @@ plot_radar_int_Shapley <- function(shapley,
 #' @param shape_label Character. Label for the shape class. If NULL (default), no legend for the shape class is shown.
 #' @param ggplotly Logical. If \code{TRUE} (default), the plot is converted to an interactive [plotly] object.
 #' @param label_obs A vector with the names of the observations to be labeled in the plot when \code{ggplotly = FALSE}. Default is NULL.
+#' @param y_limits A numeric vector of length 2 specifying the limits for the y-axis. Default is NULL, so the limits are determined automatically.
 #' @return Returns a beeswarm plot that displays the Shapley values (\code{\link{int_Shapley}}) for each observation and feature.
 #' @export
 #' 
@@ -598,7 +599,8 @@ plot_beeswarm_int_Shapley <- function(shapley,
                                  shape_class = NULL, 
                                  shape_label = NULL, 
                                  ggplotly = FALSE, 
-                                 label_obs = NULL) {
+                                 label_obs = NULL,
+                                 y_limits = NULL) {
 
   if(!is.matrix(shapley)) stop("`shapley` must be a matrix.")
 
@@ -693,6 +695,14 @@ plot_beeswarm_int_Shapley <- function(shapley,
   # Palette
   if (!is.null(palette)) {
     p <- p + scale_color_manual(values = palette)
+  }
+
+  # Y-axis limits
+  if (!is.null(y_limits)) {
+    if (!is.numeric(y_limits) || length(y_limits) != 2) {
+      stop("`y_limits` must be a numeric vector of length 2.")
+    }
+    p <- p + ylim(y_limits)
   }
 
   # Conditionally hide legends